backend/requirements.txt
asgiref==3.8.1
Django==5.0.3
django-cors-headers==4.5.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
packaging==24.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.4.4
typing_extensions==4.10.0
